requirements.txt
# Web framework
fastapi>=0.110,<0.116
uvicorn[standard]>=0.29,<0.31
streamlit>=1.33,<1.45

# Core data/runtime
numpy>=1.24.0,<1.27.0
pandas>=2.0.3,<2.1.0
pyarrow>=14.0.0,<15.0.0
xlsxwriter>=3.1,<4.0

# AI / retrieval
sentence-transformers>=2.6,<3.1
faiss-cpu>=1.7.4,<1.9

# SQL / parsing
sqlglot>=24,<26
sqlalchemy>=2.0,<2.1

# HTTP / config
httpx>=0.27,<0.29
python-dotenv>=1.0,<2.0
pyyaml>=6.0,<7.0
python-multipart>=0.0.9,<0.0.11
